ADD: uploadImage con cloudinary

+++ src/modules/daybook/views/EntryView.vue
@@ -7,6 +7,14 @@
         <span class="mx-2 fs-4 fw-light">{{ yearDay }}</span>
       </div>
       <div>
+        <input
+          type="file"
+          @change="onSelectedImage"
+          ref="imageSelector"
+          v-show="false"
+          accept="image/png, image/jpeg"
+        />
+
         <button
           v-if="entry.id"
           class="btn btn-danger mx-2"
@@ -14,7 +22,7 @@
         >
           Borrar <i class="fa fa-trash-alt"></i>
         </button>
-        <button class="btn btn-primary">
+        <button class="btn btn-primary" @click="onSelectImage">
           Subir foto <i class="fa fa-upload"></i>
         </button>
       </div>
@@ -26,7 +34,14 @@
       <textarea v-model="entry.text" placeholder="¿Qué sucedio hoy?"></textarea>
     </div>
     <img
-      src="https://enviajes.cl/wp-content/uploads/2021/04/Chile-lugares-Parinacota-y-Chungara-Enviajes.jpg"
+      v-if="entry.picture && !localImage"
+      :src="entry.picture"
+      alt="entry-picture"
+      class="img-thumbnail"
+    />
+    <img
+      v-if="localImage"
+      :src="localImage"
       alt="entry-picture"
       class="img-thumbnail"
     />
@@ -38,7 +53,9 @@
 <script>
 import { defineAsyncComponent } from "vue";
 import { mapActions, mapGetters } from "vuex";
+import Swal from "sweetalert2";
 import getDayMonthYear from "../helpers/getDayMonthYear";
+import uploadImage from "../helpers/uploadImage";
 
 export default {
   props: {
@@ -55,6 +72,8 @@ export default {
   data() {
     return {
       entry: null,
+      localImage: null,
+      file: null,
     };
   },
   computed: {
@@ -89,18 +108,61 @@ export default {
       this.entry = entry;
     },
     async saveEntry() {
+      new Swal({
+        title: "Espere por favor",
+        allowOutsideClick: false,
+      });
+      Swal.showLoading();
+
+     const picture =await uploadImage(this.file)
+
+     this.entry.picture = picture
       if (this.entry.id) {
         await this.updateEntry(this.entry);
       } else {
         const id = await this.createEntries(this.entry);
         this.$router.push({ name: "entry", params: { id } });
       }
+
+      Swal.fire("Guardando", "Entrada registrada con éxito", "success");
+      this.file = null
     },
     async onDeleteEntry() {
-      console.log("delete", this.entry);
-      await this.deleteEntry(this.entry.id);
+      const { isConfirmed } = await Swal.fire({
+        title: "¿Estás seguro?",
+        text: "Una vez borrado, no se puede recuperar",
+        showDenyButton: true,
+        confirmButtonText: "Sí, estoy seguro",
+      });
 
-      this.$router.push({ name: "no-entry" });
+      if (isConfirmed) {
+        new Swal({
+          title: "Espere por favor",
+          allowOutsideClick: false,
+        });
+        Swal.showLoading();
+
+        await this.deleteEntry(this.entry.id);
+        this.$router.push({ name: "no-entry" });
+        Swal.fire("Eliminado", "", "success");
+      }
+    },
+
+    onSelectedImage(event) {
+      const file = event.target.files[0];
+      if (!file) {
+        this.localImage = null;
+        this.file = null;
+        return;
+      }
+
+      this.file = file;
+      const fr = new FileReader();
+      fr.onload = () => (this.localImage = fr.result);
+      fr.readAsDataURL(file);
+    },
+    onSelectImage() {
+      this.$refs.imageSelector.click();
     },
   },
   created() {
